requirements.txt
feedparser
requests
python-dateutil
Pillow
beautifulsoup4
